bobinator
Struct Timeoff

pub struct Timeoff {

    pub id: i64,
    pub approved_by: Option<String>,
    pub can_edit: Option<bool>,
    pub date_range: DateRange,
    pub description: Option<String>,
    pub documents: Option<Vec<Value>>,
    pub duration: Option<i64>,
    pub duration_description: Option<String>,
    pub employee_id: String,
    pub end_date: NaiveDate,
    pub end_date_portion: DatePortion,
    pub policy_type: String,
    pub policy_type_display_name: TimeoffPolicyType,
    pub policy_type_order: i64,
    pub reason: Option<String>,
    pub reason_code_display_name: Option<String>,
    pub reason_code_id: Option<String>,
    pub request_range_type: Option<RequestRangeType>,
    pub requested_by: Option<String>,
    pub start_date: NaiveDate,
    pub start_date_portion: DatePortion,
    pub states: Option<String>,
    pub status: Option<ApprovalState>,
    pub unit: Option<String>,
    /* private fields */

}

A struct representing a single timeoff received from bob.

This struct is used for your own timeoffs as well as the reduced fields when querying other people’s timeoffs, hence the number of optional fields.

Sample return from api/timeoff/employees/{employee_id}/requests/{request_id}:

{
    "actions": [
        "editRequest",
        "cancelRequest"
    ],
    "approvedBy": null,
    "canEdit": true,
    "dateRange": {
        "startDate": "2023-11-24",
        "startPortion": "all_day",
        "endDate": "2023-11-24",
        "endPortion": "all_day",
        "type": "days"
    },
    "description": null,
    "documents": [],
    "duration": 1,
    "employeeId": "0000000000000000000",
    "endDate": "2023-11-24",
    "endDatePortion": "all_day",
    "events": [
        {
            "id": 98765432,
            "companyId": 999999,
            "requestId": 12345678,
            "eventType": "submitted",
            "eventData": {
                "request": {
                    "id": 12345678,
                    "companyId": 999999,
                    "employeeId": "0000000000000000000",
                    "dateRange": {
                        "startDate": "2023-11-24",
                        "startPortion": "all_day",
                        "endDate": "2023-11-24",
                        "endPortion": "all_day",
                        "type": "days"
                    },
                    "policyTypeReservedName": "type2",
                    "visibility": null,
                    "status": "approved",
                    "userData": {
                        "attachments": []
                    },
                    "sentTo": null
                },
                "policyApprovalRequired": false,
                "submitterApproved": false,
                "approvalRequired": false,
                "approverSchemeSource": "policy"
            },
            "origin": "web",
            "createdBy": "0000000000000000000",
            "creationTime": "2023-02-28T00:01:38.695715"
        }
    ],
    "id": 13015100,
    "latestState": {
        "tier": null,
        "status": "approved",
        "submittedAt": "2023-02-28T00:01:38.695715"
    },
    "policyType": "type2",
    "policyTypeDisplayName": "Friday Off",
    "policyTypeEmoji": "\u26d4",
    "policyTypeOrder": 5,
    "reason": null,
    "reasonCodeDisplayName": null,
    "reasonCodeId": null,
    "requestRangeType": "days",
    "requestedBy": "0000000000000000000",
    "startDate": "2023-11-24",
    "startDatePortion": "all_day",
    "states": null,
    "status": "approved",
    "unit": "days"
}

Fields
id: i64approved_by: Option<String>can_edit: Option<bool>date_range: DateRangedescription: Option<String>documents: Option<Vec<Value>>duration: Option<i64>duration_description: Option<String>employee_id: Stringend_date: NaiveDateend_date_portion: DatePortionpolicy_type: Stringpolicy_type_display_name: TimeoffPolicyTypepolicy_type_order: i64reason: Option<String>reason_code_display_name: Option<String>reason_code_id: Option<String>request_range_type: Option<RequestRangeType>requested_by: Option<String>start_date: NaiveDatestart_date_portion: DatePortionstates: Option<String>status: Option<ApprovalState>unit: Option<String>