bobinator
Trait HasDateRange

pub trait HasDateRange {
    // Required method
    fn date_range<'a>(&'a self) -> RangeInclusive<NaiveDate>;
}

A trait indicating the struct has an associated range of dates.

Required Methods

fn date_range<'a>(&'a self) -> RangeInclusive<NaiveDate>

Return the an inclusive range of dates of the instance.