Crate bobinator

Modules

app

Commmand Line App Interface Functions.

bob

consts

Structs

APIToken

An API Token, together with scope information.

APITokenOnly

Connection

Credentials

A set of login credentials to send to the server.

DEFAULT_ACCEPT

DEFAULT_ACCEPT_ENCODING

DEFAULT_ACCEPT_LANGUAGE

DEFAULT_CONNECTION

DEFAULT_ORIGIN

DEFAULT_REFERER

DEFAULT_TE

DEFAULT_USER_AGENT

DateRange

Headers

Constructs headers using the default of:

LoginSession

Parse data returned from POST /api/login endpoint.

Metadata

A “metadata” object returned by an API endpoint.

Timeoff

A struct representing a single timeoff received from bob.

TimeoffRequest

Construct a JSON for requesting time off.

UnexpectedError

The default unexpected error format returned by Bob.

WeekdayInterval

A struct grouping together staticmethods about Fridays.

WeekdayIntervalIter

Enums

APITokenScope

API token scope item.

ApprovalState

BobinatorError

DatePortion

DateRangeType

RequestRangeType

Units for the timeoff request range. This is supported only for policy types measured in hours.

TimeoffPolicyType

Policy Type for mapping the Holiday Balance.

UserInput

User Input Options.

Constants

CALENDAR_WIDTH

Traits

CalendarMonthShiftModifier

Trait for shifting the cursor to the line of the relevant date.

Capitalise

Trait to allow turning a string Capitalised.

FindDate

HasDate

A trait indicating the struct has an associated date.

HasDateRange

A trait indicating the struct has an associated range of dates.

HasToken

A trait indicating the struct can return a token key.

JoinString

TitleCase

Trait to allow turning a string Title Case.

Functions

flush_stdout

Type Definitions

FridayOff

Type Alias for Bi-weekly Friday offs.