Trait bobinator::CalendarMonthShiftModifier

pub trait CalendarMonthShiftModifier<T>where
    T: HasDate,{
    fn week_modifier(&self, n: u32) -> Option<Modifier>;
    fn shift_modifier_for(&self, obj: &T) -> Option<Modifier>;
    fn shifted_input_for<F>(&self, obj: &T, f: F) -> Option<UserInput>
    where
        F: FnOnce() -> UserInput;

    fn shift_print_on_week(&self, n: u32, text: &str) -> Option<()> { ... }
    fn shifted_print_for(&self, obj: &T, text: &str) -> Option<()> { ... }
    fn wipe_line_for(&self, obj: &T) -> Option<()> { ... }
}

Trait for shifting the cursor to the line of the relevant date.

Required Methods

fn week_modifier(&self, n: u32) -> Option<Modifier>

Return the [Modifier] that shifts the cursor to the end of line belong to week n.

This assumes that the current cursor is at the new line immediately after printing the calendar.

If the [CalendarMonth] instance does not have an nth line, return None.

fn shift_modifier_for(&self, obj: &T) -> Option<Modifier>

Return the [Modifier] that shifts the cursor to the end of line belonging to obj.

This assumes that the current cursor is at the new line immediately after printing the calendar.

If obj is not in the [CalendarMonth] instance, return None.

fn shifted_input_for<F>(&self, obj: &T, f: F) -> Option<UserInput>where
F: FnOnce() -> UserInput,

Use f to produce a user input prompt at the end of line belonging to obj, then return the UserInput.

If obj is not in the [CalendarMonth] instance, return None without printing the prompt.

Provided Methods

fn shift_print_on_week(&self, n: u32, text: &str) -> Option<()>

Assuming we are at the new line immediately after printing the calendar, shift the cursor to the nth week and print out text.

If obj is not in the [CalendarMonth] instance, return None.

fn shifted_print_for(&self, obj: &T, text: &str) -> Option<()>

Assuming we are at the new line immediately after printing the calendar, shift the cursor to the line belonging to obj and print out text.

If obj is not in the [CalendarMonth] instance, return None.

fn wipe_line_for(&self, obj: &T) -> Option<()>

Overwrite all characters on the line belonging to obj.

This assumes that the current cursor is at the new line immediately after printing the calendar.